sensor-client/requirements.txt
# AgriNex Sensor Client Dependencies v2.0

# Core dependencies
paho-mqtt>=1.6.1
pyserial>=3.5
asyncio-mqtt>=0.13.0

# Data processing
numpy>=1.21.0

# Optional dependencies (for extended features)
# opencv-python>=4.5.0  # Uncomment if camera support needed
# Pillow>=8.3.0        # Uncomment if image processing needed

# Development and testing
pytest>=6.0.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=0.19.0
psutil>=5.8.0

# 开发和测试
pytest>=6.2.0
pytest-asyncio>=0.15.0

# 可选依赖 (用于高级功能)
# 数据库支持
# sqlalchemy>=1.4.0
# pymongo>=3.12.0

# 消息队列
# redis>=3.5.0

# 可视化
# matplotlib>=3.4.0
# plotly>=5.0.0

# 配置管理
# pyyaml>=5.4.0
# python-dotenv>=0.19.0
